+++ modules/apps/google-places/google-places-web/src/main/java/com/liferay/google/places/web/internal/configuration/admin/display/GooglePlacesSiteSettingsConfigurationScreenWrapper.java
@@ -3,17 +3,19 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.google.places.web.internal.portal.settings.configuration.admin.display;
+package com.liferay.google.places.web.internal.configuration.admin.display;
 
 import com.liferay.configuration.admin.display.ConfigurationScreen;
 import com.liferay.configuration.admin.display.ConfigurationScreenWrapper;
 import com.liferay.google.places.constants.GooglePlacesWebKeys;
 import com.liferay.google.places.util.GooglePlacesUtil;
 import com.liferay.portal.kernel.language.Language;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenContributor;
-import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenFactory;
+import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenContributor;
+import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenFactory;
 
 import java.util.Locale;
 
@@ -28,27 +30,30 @@ import org.osgi.service.component.annotations.Reference;
  * @author Rodrigo Paulino
  */
 @Component(service = ConfigurationScreen.class)
-public class GooglePlacesPortalSettingsConfigurationScreenWrapper
+public class GooglePlacesSiteSettingsConfigurationScreenWrapper
 	extends ConfigurationScreenWrapper {
 
 	@Override
 	protected ConfigurationScreen getConfigurationScreen() {
-		return _portalSettingsConfigurationScreenFactory.create(
-			new GooglePlacesPortalSettingsConfigurationScreenContributor());
+		return _siteSettingsConfigurationScreenFactory.create(
+			new GooglePlacesSiteSettingsConfigurationScreenContributor());
 	}
+
+	@Reference
+	private GroupLocalService _groupLocalService;
 
 	@Reference
 	private Language _language;
 
-	@Reference
-	private PortalSettingsConfigurationScreenFactory
-		_portalSettingsConfigurationScreenFactory;
-
 	@Reference(target = "(osgi.web.symbolicname=com.liferay.google.places.web)")
 	private ServletContext _servletContext;
 
-	private class GooglePlacesPortalSettingsConfigurationScreenContributor
-		implements PortalSettingsConfigurationScreenContributor {
+	@Reference
+	private SiteSettingsConfigurationScreenFactory
+		_siteSettingsConfigurationScreenFactory;
+
+	private class GooglePlacesSiteSettingsConfigurationScreenContributor
+		implements SiteSettingsConfigurationScreenContributor {
 
 		@Override
 		public String getCategoryKey() {
@@ -57,22 +62,17 @@ public class GooglePlacesPortalSettingsConfigurationScreenWrapper
 
 		@Override
 		public String getJspPath() {
-			return "/portal_settings/google_places.jsp";
+			return "/site_settings/google_places.jsp";
 		}
 
 		@Override
 		public String getKey() {
-			return "google-places-portal-settings";
+			return "google-places-site-settings";
 		}
 
 		@Override
 		public String getName(Locale locale) {
 			return _language.get(locale, "google-places");
-		}
-
-		@Override
-		public String getSaveMVCActionCommandName() {
-			return "/portal_settings/edit_company";
 		}
 
 		@Override
@@ -85,17 +85,20 @@ public class GooglePlacesPortalSettingsConfigurationScreenWrapper
 			HttpServletRequest httpServletRequest,
 			HttpServletResponse httpServletResponse) {
 
-			PortalSettingsConfigurationScreenContributor.super.setAttributes(
+			SiteSettingsConfigurationScreenContributor.super.setAttributes(
 				httpServletRequest, httpServletResponse);
 
 			ThemeDisplay themeDisplay =
 				(ThemeDisplay)httpServletRequest.getAttribute(
 					WebKeys.THEME_DISPLAY);
 
+			Group group = themeDisplay.getSiteGroup();
+
 			httpServletRequest.setAttribute(
 				GooglePlacesWebKeys.GOOGLE_PLACES_API_KEY,
 				GooglePlacesUtil.getGooglePlacesAPIKey(
-					themeDisplay.getCompanyId()));
+					themeDisplay.getCompanyId(), group.getGroupId(),
+					_groupLocalService));
 		}
 
 	}
